+++ docs/guide/demos/tokens.ts
@@ -1,9 +1,17 @@
-import type { ThemeConfig } from 'antd';
+import type { Web3ThemeConfig } from '@ant-design/web3';
 import { theme } from 'antd';
 
-export type ThemeValue = 'default' | 'violet' | 'dark' | 'green' | 'pink';
+export type ThemeValue =
+  | 'default'
+  | 'violet'
+  | 'dark'
+  | 'green'
+  | 'pink'
+  | 'retro'
+  | 'win95'
+  | 'material';
 
-export const customToken: ThemeConfig = {
+export const customToken: Web3ThemeConfig = {
   token: {
     borderRadius: 16,
     wireframe: false,
@@ -64,7 +72,8 @@ export type ThemeSetting = {
   color: string;
   value: ThemeValue;
   name: string;
-  token: ThemeConfig;
+  token: Web3ThemeConfig;
+  buttonType?: 'primary' | 'dashed' | 'link' | 'text' | 'default';
 };
 
 // 部分参考 antd 官网的主题 https://github.com/ant-design/ant-design/blob/master/.dumi/pages/index/components/Theme/index.tsx#L305
@@ -96,6 +105,98 @@ export const themeList: ThemeSetting[] = [
         },
       },
       algorithm: theme.darkAlgorithm,
+    },
+  },
+  {
+    color: '#008080',
+    value: 'win95',
+    name: 'Windows 95',
+    buttonType: 'default',
+    token: {
+      token: {
+        colorPrimary: '#008080',
+        colorLink: '#0000FF',
+        colorBgContainer: '#C0C0C0',
+        borderRadius: 0,
+      },
+      components: {
+        Button: {
+          defaultBg: '#C0C0C0',
+          defaultBorderColor: '#FFFFFF #808080 #808080 #FFFFFF',
+          defaultHoverBg: '#C0C0C0',
+          defaultShadow: 'inset -1px -1px #0a0a0a,inset 1px 1px #fff',
+        },
+        Modal: {
+          contentBg: '#C0C0C0',
+          boxShadow: 'inset -1px -1px #0a0a0a,inset 1px 1px #fff',
+        },
+      },
+      web3Components: {
+        ConnectModal: {
+          hoverWalletBg: '#DFDFDF',
+        },
+      },
+    },
+  },
+  {
+    color: '#796d6f',
+    value: 'retro',
+    name: 'Retro',
+    buttonType: 'default',
+    token: {
+      token: {
+        colorPrimary: '#000000',
+        colorLink: '#8b837d',
+        colorBgContainer: '#f3eae4',
+      },
+      components: {
+        Button: {
+          defaultShadow: '-4px 4px 0px #000000,inset 0 0 0 2px #000000',
+          defaultHoverBg: '#f3eae4',
+          defaultBg: '#fff',
+          defaultBorderColor: '#000000',
+        },
+        Modal: {
+          boxShadow: '-10px 10px 0px #000000,inset 0 0 0 2px #000000',
+          contentBg: '#eadcd1',
+        },
+      },
+      web3Components: {
+        ConnectModal: {
+          hoverWalletBg: '#f3eae4',
+        },
+      },
+    },
+  },
+  {
+    color: '#6750A4',
+    value: 'material',
+    name: 'Material Design 3',
+    buttonType: 'primary',
+    token: {
+      token: {
+        colorPrimary: '#6750a4',
+        colorBorderBg: '',
+        colorLink: '#6750a4',
+        colorBgBase: '#fdfbff',
+        colorBgContainer: '#fdfbff',
+        colorBgMask: 'rgba(0, 0, 0, 0.32)',
+        controlHeight: 40,
+      },
+      components: {
+        Button: {
+          colorBorder: 'transparent',
+          colorText: '#6750a4',
+          colorPrimaryBg: '#6750a4',
+          primaryShadow: 'none',
+          dangerShadow: 'none',
+          defaultShadow: 'none',
+        },
+        Modal: {
+          contentBg: '#fffbfe',
+          boxShadow: 'none',
+        },
+      },
     },
   },
   {