+++ packages/icons/src/__tests__/iconTest.tsx
@@ -1,8 +1,9 @@
-import path from 'path';
+import path from 'node:path';
 import { type ComponentType } from 'react';
 import { globSync } from 'glob';
 
 export async function getComponents(dir: string): Promise<ComponentType[]> {
+  // biome-ignore lint/suspicious/noAsyncPromiseExecutor: just for test
   return new Promise(async (resolve) => {
     const files = globSync(path.resolve(__dirname, '..', dir, '*.tsx'));
     const res: ComponentType[] = [];