+++ packages/web3/src/crypto-input/demos/swapMode.tsx
@@ -47,6 +47,7 @@ const App: React.FC = () => {
         }}
         options={[ETH, USDT]}
       />
+      {/* biome-ignore lint/a11y/useKeyWithClickEvents: <explanation> */}
       <span
         style={{
           width: 30,